Refactor optimisation._nlopt module into directory
Instead of one big file.
